+++ src/preview.ts
@@ -2,6 +2,7 @@ import { NumberClass } from './index.ts';
 import './preview.css';
 
 document.querySelector<HTMLDivElement>('#app')!.innerHTML = `
+  <div id="key-pressed"></div>
   <div class="main-form">
     <div id="result"></div>
     <input
@@ -17,9 +18,10 @@ document.querySelector<HTMLDivElement>('#app')!.innerHTML = `
   </div>
 `;
 
-const inp = document.querySelector('#num');
+const inp: HTMLInputElement | null = document.querySelector('#num');
 const nc = new NumberClass(inp);
 const result = document.querySelector('#result');
+const keyPressed = document.querySelector('#key-pressed');
 
 function generateRandomNumber(min: number, max: number) {
   return Math.floor(Math.random() * (max - min + 1)) + min;
@@ -33,13 +35,19 @@ const buttonClicked = () => {
   }
 };
 
-buttonClicked();
+// buttonClicked();
 
 nc.addEventListener('input', (evt) => {
   if (result) {
     result.innerHTML = evt.detail.value || '';
   }
 });
+
+if (inp) {
+  inp.addEventListener('keydown', (evt: KeyboardEvent) => {
+    if (keyPressed) keyPressed.innerHTML = evt.key;
+  });
+}
 
 const randButton = document.querySelector('#rand-button');
 if (randButton) {